Refactor toward testability by introducing Editor
Jira ticket: CAMS-526
